requirements.txt
Flask
psycopg2-binary
python-dotenv
google-cloud-language
pandas
Flask-Bcrypt
Flask-Mail
fpdf
gunicorn
